account login CUD extended function & manytomany relation

+++ 03_Django/06_django_form/articles/views.py
@@ -3,10 +3,10 @@ from django.views.decorators.http import require_POST
 from django.contrib.auth.decorators import login_required
 from .models import Article, Comment
 from .forms import ArticleForm, CommentForm
+import hashlib
 from IPython import embed
 
 def index(request):
-    # embed()
     # 1. session 정보에서 visits_num 이라는 키로 접근해 값을 가져옴
     # 해당하는 키가 없으면 0을 가져옴
     visits_num = request.session.get('visits_num', 0)
@@ -38,7 +38,10 @@ def create(request):
         form = ArticleForm(request.POST)
         # 해당 폼이 유효한지 확인
         if form.is_valid():
-            article = form.save()
+            article = form.save(commit=False)
+            article.user_id = request.user.id
+            article.save()
+            embed()
             return redirect('articles:detail', article.pk)
     else:
         form = ArticleForm()
@@ -57,26 +60,29 @@ def detail(request, article_pk):
 def delete(request, article_pk):
     if request.user.is_authenticated:
         article = get_object_or_404(Article, pk=article_pk)
-        article.delete()
+        if request.user.id == article.user.id:
+            article.delete()
     return redirect('articles:index') # redirect -> GET 요청
 
 @login_required
 def update(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
-    if request.method == 'POST':
-        # instance -> 수정의 대상이 되는 특정한 글 객체
-        form = ArticleForm(request.POST, instance=article)
-        if form.is_valid():
-            form.save()
-            # embed()
-            return redirect('articles:detail', article.pk)
+    if request.user.id == article.user.id:
+        if request.method == 'POST':
+            # instance -> 수정의 대상이 되는 특정한 글 객체
+            form = ArticleForm(request.POST, instance=article)
+            if form.is_valid():
+                form.save()
+                return redirect('articles:detail', article.pk)
+        else:
+            # form = ArticleForm(initial={
+            #     'title': article.title,
+            #     'content': article.content,
+            # })
+            form = ArticleForm(instance=article)
     else:
-        # form = ArticleForm(initial={
-        #     'title': article.title,
-        #     'content': article.content,
-        # })
-        form = ArticleForm(instance=article)
-        # embed()
+        return redirect('articles:index')
+
     context = {'form': form, 'article': article, }
     return render(request, 'articles/form.html', context)
 
@@ -103,6 +109,7 @@ def comments_create(request, article_pk):
         if comment_form.is_valid():
             comment = comment_form.save(commit=False)
             comment.article_id = article_pk
+            comment.user_id = request.user.id
             comment.save()
     return redirect('articles:detail', article_pk)
 
@@ -111,5 +118,6 @@ def comments_create(request, article_pk):
 def comments_delete(request, article_pk, comment_pk):
     if request.user.is_authenticated:    
         comment = get_object_or_404(Comment, pk=comment_pk)
-        comment.delete()
+        if comment.user == request.user:
+            comment.delete()
     return redirect('articles:detail', article_pk)
